requirements.txt
APScheduler==3.6.3
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
decorator==4.4.2
Flask==1.1.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
service-identity==18.1.0
six==1.15.0
toml==0.10.2
tornado==6.1
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.2
waitress==1.4.4
Werkzeug==1.0.1
wincertstore==0.2
zope.interface==5.2.0
